Fuck you
The teacher is never in office hours. I am too busy to do everything
she asks. It just feels as if the workload does not acknowledge that I
have other commitments to do. It’s a selfish child that does not give a
shit about the other subjects I need to attend to.